fix: 상품 관련 로직 중 레이어드 아키텍처 위반한 부분 수정, 가격 타입 BigDecimal로 통일

+++ src/main/java/com/hhplusecommerce/domain/product/Product.java
@@ -7,6 +7,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 /**
@@ -21,12 +22,12 @@ public class Product {
     private Long id;
     private String name;
     private String category;
-    private Long price;
+    private BigDecimal price;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
 
     @Builder
-    public Product(Long id, String name, String category, Long price) {
+    public Product(Long id, String name, String category, BigDecimal price) {
         this.id = id;
         this.name = name;
         this.category = category;
